+++ src/nmrt/core/segment.py
@@ -8,8 +8,8 @@ from .model import Recording, Signal
 def segment_signal(signal: Signal, start_s: float, end_s: float, *, name: str | None = None) -> Signal:
     if end_s <= start_s:
         raise ValueError("end_s must exceed start_s")
-    i0 = max(0, int(round((start_s - signal.start_time_s) * signal.fs)))
-    i1 = min(signal.n_samples, int(round((end_s - signal.start_time_s) * signal.fs)))
+    i0 = max(0, round((start_s - signal.start_time_s) * signal.fs))
+    i1 = min(signal.n_samples, round((end_s - signal.start_time_s) * signal.fs))
     if i1 <= i0:
         raise ValueError("Requested segment does not overlap signal")
     return replace(signal, name=name or signal.name, data=signal.data[i0:i1], start_time_s=start_s)
